Add files via upload
Hello! I developed this dashboard from a sample dataset within the Power BI. This dashboard have three pages:

01. Home Page
This is for presentation of dashboard and to understand what you are about to see, pretty simple.

02. Overview Analytics
This, as the name suggers, present an overview of the business with KPI's, like total of sales, profit, discounts and more.

03. Financial Analytics
For last, an analysis of the profit by segment, gross sales and all related to this!